moving vocab to legacy folder
