+++ examples/BERT/qa_task.py
@@ -163,7 +163,7 @@ if __name__ == "__main__":
     except:
         train_dataset, dev_dataset = SQuAD1()
         old_vocab = train_dataset.vocab
-        vocab = torchtext.vocab.Vocab(counter=old_vocab.freqs,
+        vocab = torchtext.legacy.vocab.Vocab(counter=old_vocab.freqs,
                                       specials=['<unk>', '<pad>', '<MASK>'])
         with open(args.save_vocab, 'wb') as f:
             torch.save(vocab, f)
