+++ examples/vocab/vocab.py
@@ -4,7 +4,7 @@ import argparse
 import torch
 import io
 
-from torchtext.vocab import build_vocab_from_iterator
+from torchtext.legacy.vocab import build_vocab_from_iterator
 from torchtext.data.utils import ngrams_iterator
 from torchtext.data.utils import get_tokenizer
 from torchtext.utils import unicode_csv_reader
