+++ test/legacy/sst.py
@@ -1,6 +1,6 @@
 from torchtext.legacy import data
 from torchtext.legacy import datasets
-from torchtext.vocab import Vectors, GloVe, CharNGram, FastText
+from torchtext.legacy.vocab import Vectors, GloVe, CharNGram, FastText
 
 
 # Approach 1:
