+++ torchtext/legacy/__init__.py
@@ -2,7 +2,7 @@ from . import data
 from .. import nn  # Not in the legacy folder
 from . import datasets
 from .. import utils  # Not in the legacy folder
-from .. import vocab  # Not in the legacy folder
+from . import vocab
 
 __all__ = ['data',
            'nn',
