+++ torchtext/legacy/datasets/text_classification.py
@@ -4,8 +4,8 @@ import io
 from torchtext.utils import download_from_url, extract_archive, unicode_csv_reader
 from torchtext.data.utils import ngrams_iterator
 from torchtext.data.utils import get_tokenizer
-from torchtext.vocab import build_vocab_from_iterator
-from torchtext.vocab import Vocab
+from torchtext.legacy.vocab import build_vocab_from_iterator
+from torchtext.legacy.vocab import Vocab
 from tqdm import tqdm
 
 URLS = {
